+++ src/main/java/com/dnsimple/endpoints/Billing.java
@@ -23,6 +23,6 @@ public class Billing {
      * @see <a href="https://developer.dnsimple.com/v2/billing-charges/#listCharges">https://developer.dnsimple.com/v2/billing-charges/#listCharges</a>
      */
     public ListResponse<Charge> listCharges(Number account, ListOptions options) {
-        return client.list(GET, String.format("/%s/billing/charges", account), options, null, Charge.class);
+        return client.list(GET, account + "/billing/charges", options, null, Charge.class);
     }
 }